requirements.txt
pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
